drm/i915: Handle validation of relocation object with size = 2GB
With prefaulting extended to support 2GB size, i915 can handle
relocation objects of 2GB.

v2: Removed relocs_maxs and relocs_count type change.

Testcase: igt/gem_exec_reloc/wc-31
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Cc: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Joonas Lahtinen <joonas.lahtinen@linux.intel.com>

+++ drivers/gpu/drm/i915/i915_gem_execbuffer.c
@@ -1154,7 +1154,7 @@ validate_exec_list(struct drm_device *dev,
 
 	for (i = 0; i < count; i++) {
 		char __user *ptr = u64_to_user_ptr(exec[i].relocs_ptr);
-		int length; /* limited by fault_in_pages_readable() */
+		size_t length;
 
 		if (exec[i].flags & invalid_flags)
 			return -EINVAL;
